test: 입력 받는 테스트 파일과 당첨 결과 테스트 파일 분리
